backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
langchain
langchain-google-genai
faiss-cpu
numpy
pypdf
pdfminer.six
python-multipart
requests
